+++ apps/posts/src/views/PostAnalytics/PostAnalytics.tsx
@@ -1,8 +1,11 @@
+import PostAnalyticsLayout from './components/layout/PostAnalyticsLayout';
 import {Outlet} from '@tryghost/admin-x-framework';
 
 const PostAnalytics: React.FC = () => {
     return (
-        <Outlet />
+        <PostAnalyticsLayout>
+            <Outlet />
+        </PostAnalyticsLayout>
     );
 };
